fix: switch dist files to use babili, both be es5
The `idbify.min.js` was set as the `module` key - except that it used
ES6 syntaxes beyond `export`.

The `module` key is defined as using only ES6 module syntax
(`import`,`export`) and _not_ any additional syntax, beyond the baseline
browser support. This means for ES5 browsers (such as IE11) the UMD
version would need to be used over the Modules version because the UMD
was the only one with ES5 support.

This change makes both the `module` version and the `umd` versions ES5
syntax only. In addition it gives them more accurate names - the
`module` version goes fro `min` to `module` and the `umd` version goes
from `es5` to `umd` - this more accurately reflects the output of these
formats, which are now both ES5 syntax, the only difference being the
module format.

Lastly, the scripts have been changed subtly - so that `npm run pack`
now becomes `npm run prepare` - this is a lifecycle hook that runs when
`npm pack` or `npm publish` runs - negating the need to run the
additional script manually.